requirements.txt
streamlit==1.29.0
tensorflow-cpu==2.15.0
numpy==1.26.4
Pillow==10.4.0
google-generativeai==0.8.6
protobuf==4.25.3
